ansible: load watches in run
With the watches package, we no longer need operator.go for ansible
based operators.
