+++ pkg/ansible/run.go
@@ -19,11 +19,15 @@ import (
 	"fmt"
 	"os"
 	"runtime"
+	"strconv"
+	"strings"
 
+	"github.com/operator-framework/operator-sdk/pkg/ansible/controller"
 	aoflags "github.com/operator-framework/operator-sdk/pkg/ansible/flags"
-	"github.com/operator-framework/operator-sdk/pkg/ansible/operator"
 	proxy "github.com/operator-framework/operator-sdk/pkg/ansible/proxy"
 	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/controllermap"
+	"github.com/operator-framework/operator-sdk/pkg/ansible/runner"
+	"github.com/operator-framework/operator-sdk/pkg/ansible/watches"
 	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 	"github.com/operator-framework/operator-sdk/pkg/leader"
 	"github.com/operator-framework/operator-sdk/pkg/metrics"
@@ -31,11 +35,13 @@ import (
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
+	"sigs.k8s.io/controller-runtime/pkg/runtime/signals"
 )
 
 var (
@@ -80,13 +86,45 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 		return err
 	}
 
-	name, found := os.LookupEnv(k8sutil.OperatorNameEnvVar)
-	if !found {
-		log.Error(fmt.Errorf("%s environment variable not set", k8sutil.OperatorNameEnvVar), "")
+	cMap := controllermap.NewControllerMap()
+	watches, err := watches.Load(flags.WatchesFile)
+	if err != nil {
+		log.Error(err, "Failed to load watches.")
+	}
+	for _, w := range watches {
+		runner, err := runner.New(w)
+		if err != nil {
+			log.Error(err, "Failed to create runner")
+			return err
+		}
+
+		ctr := controller.Add(mgr, controller.Options{
+			GVK:             w.GroupVersionKind,
+			Runner:          runner,
+			ManageStatus:    w.ManageStatus,
+			MaxWorkers:      getMaxWorkers(w.GroupVersionKind, flags.MaxWorkers),
+			ReconcilePeriod: w.ReconcilePeriod,
+		})
+		if ctr == nil {
+			return fmt.Errorf("failed to add controller for GVK %v", w.GroupVersionKind.String())
+		}
+
+		cMap.Store(w.GroupVersionKind, &controllermap.Contents{Controller: *ctr,
+			WatchDependentResources:     w.WatchDependentResources,
+			WatchClusterScopedResources: w.WatchClusterScopedResources,
+			OwnerWatchMap:               controllermap.NewWatchMap(),
+			AnnotationWatchMap:          controllermap.NewWatchMap(),
+		})
+	}
+
+	operatorName, err := k8sutil.GetOperatorName()
+	if err != nil {
+		log.Error(err, "Failed to get the operator name")
 		return err
 	}
+
 	// Become the leader before proceeding
-	err = leader.Become(context.TODO(), name+"-lock")
+	err = leader.Become(context.TODO(), operatorName+"-lock")
 	if err != nil {
 		log.Error(err, "Failed to become leader.")
 		return err
@@ -105,7 +143,6 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 	}
 
 	done := make(chan error)
-	cMap := controllermap.NewControllerMap()
 
 	// start the proxy
 	err = proxy.Run(done, proxy.Options{
@@ -124,7 +161,9 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 	}
 
 	// start the operator
-	go operator.Run(done, mgr, flags, cMap)
+	go func() {
+		done <- mgr.Start(signals.SetupSignalHandler())
+	}()
 
 	// wait for either to finish
 	err = <-done
@@ -134,4 +173,29 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 	}
 	log.Info("Exiting.")
 	return nil
+}
+
+// if the WORKER_* environment variable is set, use that value.
+// Otherwise, use the value from the CLI. This is definitely
+// counter-intuitive but it allows the operator admin adjust the
+// number of workers based on their cluster resources. While the
+// author may use the CLI option to specify a suggested
+// configuration for the operator.
+func getMaxWorkers(gvk schema.GroupVersionKind, defValue int) int {
+	envVar := strings.ToUpper(strings.Replace(
+		fmt.Sprintf("WORKER_%s_%s", gvk.Kind, gvk.Group),
+		".",
+		"_",
+		-1,
+	))
+	switch maxWorkers, err := strconv.Atoi(os.Getenv(envVar)); {
+	case maxWorkers <= 1:
+		return 1
+	case err != nil:
+		// we don't care why we couldn't parse it just use default
+		log.Info("Failed to parse %v from environment. Using default %v", envVar, defValue)
+		return defValue
+	default:
+		return maxWorkers
+	}
 }
